feat: :white_check_mark: Create currency feature

+++ priv/repo/migrations/20220718193959_create_transactions_table.exs
@@ -4,10 +4,10 @@ defmodule Core.Repo.Migrations.CreateTransactions do
   def change do
     create table("transactions", primary_key: false) do
       add(:id, :binary_id, primary_key: true)
-      add(:origin_value, :integer)
+      add(:origin_value, :float)
       add(:origin_currency, :string)
       add(:final_currency, :string)
-      add(:convertion_tax, :integer)
+      add(:convertion_tax, :float)
       add(:user_id, references("users", type: :binary_id))
 
       timestamps(type: :utc_datetime_usec)
